Split what are now thunk-specific functions out of actions file
