+++ src/App.js
@@ -6,8 +6,9 @@ import './App.css';
 import { connect } from 'react-redux'
 import AccessionTable from 'components/Accession/AccessionTable';
 import SearchBar from 'components/SearchBar/SearchBar';
+import * as CONSTANTS from './constants';
 
-const sideEffectLibs = ['thunk', 'saga', 'epic'];
+const sideEffectLibs = [CONSTANTS.SIDE_EFFECT_LIB_THUNK, CONSTANTS.SIDE_EFFECT_LIB_SAGA, CONSTANTS.SIDE_EFFECT_LIB_EPIC];
 
 class App extends React.Component {
   render() {
