+++ src/components/SearchBar/SearchBar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import './SearchBar.css';
-import { makeQueryIfNeeded } from 'redux/actions';
+import { makeQueryIfNeeded } from 'redux/thunk';
 
 const SEARCH_BAR_ID = 'search-bar-input';
