+++ src/redux/actions.js
@@ -1,6 +1,6 @@
 import * as CONSTANTS from '../constants';
 
-export function sendSearchQuery(query, sideEffectLib) {
+export function requestSearchResults(query, sideEffectLib) {
   return {
     type: CONSTANTS.REQUEST_SEARCH_RESULTS,
     query,
@@ -17,16 +17,6 @@ export function receiveSearchResults(query, json, sideEffectLib) {
   };
 };
 
-function makeQuery(query) {
-  return (dispatch) => {
-    dispatch(sendSearchQuery(query, CONSTANTS.SIDE_EFFECT_LIB_THUNK));
-    // TODO: Abstract URL creation to util function
-    return fetch(`${CONSTANTS.CORS_ANYWHERE_LOCAL_URL}/${CONSTANTS.GENE_LAB_API_URL}?type=cgene&api_key=${CONSTANTS.API_KEY}&term=${query}`)
-      .then((response) => response.json())
-      .then((json) => dispatch(receiveSearchResults(query, json, CONSTANTS.SIDE_EFFECT_LIB_THUNK)));
-  };
-}
-
 export function shouldMakeQuery({stateQuery, newQuery, queryInProgress = false}) {
   if (!newQuery || newQuery === stateQuery) {
     return false;
@@ -37,18 +27,3 @@ export function shouldMakeQuery({stateQuery, newQuery, queryInProgress = false})
 
   return true;
 }
-
-export function makeQueryIfNeeded(query) {
-  return (dispatch, getState) => {
-    const state = getState();
-    const params = {
-      stateQuery: state.query,
-      newQuery: query,
-      queryInProgress: state.queryInProgress,
-    };
-
-    if (shouldMakeQuery(params)) {
-      return dispatch(makeQuery(query));
-    }
-  };
-};
